+++ lib/router.js
@@ -1751,7 +1751,7 @@ router.get('/potplayer/update/:language?', lazyloadRouteHandler('./routes/potpla
 // router.get('/caus/:category?', lazyloadRouteHandler('./routes/caus'));
 
 // 摩点
-router.get('/modian/zhongchou/:category?/:sort?/:status?', lazyloadRouteHandler('./routes/modian/zhongchou'));
+// router.get('/modian/zhongchou/:category?/:sort?/:status?', lazyloadRouteHandler('./routes/modian/zhongchou'));
 
 // 世界计划 多彩舞台 feat.初音未来 (ProjectSekai)
 // router.get('/pjsk/news', lazyloadRouteHandler('./routes/pjsk/news'));
