+++ lib/routes/modian/namespace.ts
@@ -1,0 +1,6 @@
+import type { Namespace } from '@/types';
+
+export const namespace: Namespace = {
+    name: '摩点',
+    url: 'modian.com',
+};
